Configuring middleware, propably chaning to jose, jwt is not supported in nextjs.
