+++ package.json
@@ -28,6 +28,7 @@
     "crypto-js": "^4.2.0",
     "dotenv": "^16.3.1",
     "drizzle-orm": "^0.29.3",
+    "jose": "^5.2.3",
     "jsonwebtoken": "^9.0.2",
     "lucide-react": "^0.292.0",
     "mysql2": "^3.9.1",
