+++ src/app/api/v1/auth/login/route.ts
@@ -44,7 +44,6 @@ export async function POST(request: Request) {
     });
     return NextResponse.json({
       message: "User logged in properly",
-      status: 200,
     });
   } catch (e) {
     return NextResponse.json(
